move RemoveToken
